backend/requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1

# LangGraph / LangChain ecosystem (compatible range)
langgraph>=0.2.60,<1.0
langgraph-checkpoint-postgres>=2.0.9,<3.0
langchain>=0.3.14,<0.4
langchain-core>=0.3.33,<0.4
langchain-community>=0.3.14,<0.4
langchain-groq>=0.2.4,<1.0
langchain-mcp-adapters>=0.1.1,<0.2.0
langsmith>=0.2.10,<1.0

# Guardrails (requires C++ build tools on Windows — install separately)
# nemoguardrails==0.11.0

# Database
asyncpg==0.30.0
psycopg[binary]==3.2.4
sqlalchemy==2.0.36

# Queue
redis==5.2.1

# HTTP / WebSockets
httpx==0.28.1
websockets==14.1

# Dev tools
pytest==8.3.4
pytest-asyncio==0.24.0
mypy==1.13.0
ruff==0.8.6
